fixed
fix:
changing paths of contacts path,
database
separation of model/contacts
