+++ helpers/contactsPath.js
@@ -1,0 +1,5 @@
+const path = require('path')
+
+const contactsPath = path.resolve(__dirname, '../', 'db', 'contacts.json')
+
+module.exports = contactsPath;
