+++ model/contacts/updateContacts.js
@@ -1,0 +1,8 @@
+const fs = require('fs/promises')
+const contactsPath = require('../../helpers/contactsPath');
+
+const updateContacts = async (newContact) => {
+  await fs.writeFile(contactsPath, JSON.stringify(newContact));
+};
+
+module.exports = updateContacts;
