--- model/contactsPath.js
@@ -1,5 +1,0 @@
-const path = require('path')
-
-const contactsPath = path.resolve(__dirname, 'contacts.json')
-
-module.exports = contactsPath;
